% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_mpnst_example}
\alias{sce_mpnst_example}
\title{Example MPNST SingleCellExperiment for the scOverlay vignette}
\format{
A \linkS4class{SingleCellExperiment} object with selected genes, cell
metadata and reduced dimensions.
}
\source{
European Genome-phenome Archive study EGAS50000001747,
\url{https://ega-archive.org/studies/EGAS50000001747}.
}
\description{
A small \linkS4class{SingleCellExperiment} object derived from the single-cell
MPNST progression dataset deposited at the European Genome-phenome Archive
under accession EGAS50000001747. The object has been reduced in size and is
included only to demonstrate scOverlay functionality in the package vignette.
}
\examples{
data("sce_mpnst_example")
sce_mpnst_example

}
\keyword{datasets}
