% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_overlay_example}
\alias{sce_overlay_example}
\title{Small example SingleCellExperiment for scOverlay}
\format{
A \linkS4class{SingleCellExperiment} object with gene expression
values, cell metadata and reduced dimensions.
}
\description{
A small \linkS4class{SingleCellExperiment} object used in scOverlay examples
and tests. It contains a reduced sintetic set of cells, genes, cell annotations and
reduced dimensions to demonstrate the basic plotting functions quickly.
It includes \code{logcounts}, \code{TSNE} and \code{UMAP} reduced dimensions,
and cell metadata such as \code{sample}, \code{sample_type} and
\code{cluster}.
}
\examples{
data("sce_overlay_example")
sce_overlay_example

}
\keyword{datasets}
